requirements.txt
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
commonmark==0.9.1
docutils==0.18.1
e6xdb==0.0.27
future==0.18.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
keyring==23.5.1
pkginfo==1.8.2
pycryptodome==3.14.1
Pygments==2.12.0
python-dateutil==2.8.2
pytz
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
six==1.16.0
SQLAlchemy==1.4.36
thrift==0.16.0
twine==4.0.0
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.0
